Csh16

+++ htaProject/src/controller/PurchaseController.java
@@ -8,13 +8,12 @@ import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import org.springframework.web.servlet.view.RedirectView;
 
 import model.CartDAO;
 import model.ProductDAO;
@@ -37,7 +36,7 @@ public class PurchaseController {
 	@GetMapping(value = "/final")
 	public ModelAndView purchaseOne(@RequestParam(value = "p_id") String productId,
 			@RequestParam(value = "p_count") int count, @RequestParam(value = "u_id") String userId,
-			HttpSession session, Model model) {
+			HttpSession session, Model model) throws Exception {
 
 		ModelAndView mv = new ModelAndView();
 		if (session.getAttribute("userId") == null) {
@@ -63,7 +62,7 @@ public class PurchaseController {
 	@PostMapping(value = "/final2")
 	public ModelAndView purchaseCart(@RequestParam(value = "p_id") List<String> productId,
 			@RequestParam(value = "p_count") List<String> productCount, @RequestParam(value = "cNum") List<String> cNum,
-			@RequestParam(value = "u_id") String userId) {
+			@RequestParam(value = "u_id") String userId) throws Exception {
 
 		ModelAndView mv = new ModelAndView();
 		User user = userDAO.selectOneUser(userId);
@@ -85,22 +84,41 @@ public class PurchaseController {
 
 	@PostMapping(value = "/receipt")
 	public ModelAndView receipt(@RequestParam(value = "name") String name, @RequestParam(value = "tel") String tel,
-			@RequestParam(value = "address") String address, @RequestParam(value = "cNum") List<String> cartNum) {
+			@RequestParam(value = "address") String address, @RequestParam(value = "cNum") List<String> cartNum,
+			@RequestParam(value = "p_id") String productNum, @RequestParam(value = "count") String count) throws Exception {
 		ModelAndView mv = new ModelAndView();
-		
+
 		mv.addObject("name", name);
 		mv.addObject("tel", tel);
 		mv.addObject("address", address);
 		
-		List<Cart> cartList = new ArrayList<Cart>();
-		for(String s : cartNum) {
-			cartList.add(cartDAO.selectCartNum(Integer.parseInt(s)));
-			cartDAO.deleteCart(s);
+		if(cartNum.size() == 0) {
+			mv.addObject("product", productDAO.getOneProduct(productNum));
+			mv.addObject("count", count);
+			mv.addObject("type", "one");
+		} else {
+			List<Cart> cartList = new ArrayList<Cart>();
+			for (String s : cartNum) {
+				cartList.add(cartDAO.selectCartNum(Integer.parseInt(s)));
+				cartDAO.deleteCart(s);
+			}
+			mv.addObject("receipt", cartList);
+			mv.addObject("type", "cart");
 		}
-		mv.addObject("cartList", cartList);
-		
+
 		mv.setViewName("receipt");
-		
+
 		return mv;
 	}
+
+	@ExceptionHandler(Exception.class)
+	public ModelAndView handleException(Exception e) {
+		ModelAndView mv = new ModelAndView();
+
+		mv.addObject("errorMessage", e.getMessage());
+		mv.setViewName("error");
+
+		return mv;
+	}
+
 }
